Move directory check
